Automatically connect to bluetooth on UART1
On some Overo COMs, the bluetooth module connected on ttyO1. This
hardware must be explicitly hciattach'ed---udev doesn't know about
it.  This bbappend adds a systemd service to the bluez4 package
so the adapter is available, if it exists, to the bluez stack. If the
adapter is not present, this service will harmlessly fail.