Automatically start Bluetooth Service if present.
The systemd bluetooth.service is disabled by default and looks to
be activated by d-bus. There is no activity though to effectively call
'hciconfig hci0 up' on boot so for users, bluetooth is available but
not turned on.  This patch enables the service by default turning
bluetooth on for boards that use it.