[systemd] Use systemd-networkd service to manage connections
Ifupdown tools do not work with newer kernels such as 3.17. As these
tools are deemed deprecated anyways, we should give the systemd's
network management service a try.

It still relies on wpa-supplicant for roaming and connecting to wireless
networks, but we no longer need udhcpc systemd service.

After specifying wireless network in
/etc/wpa_supplicant/wpa_supplicant-{mlan0|wlan0}.conf, simply enable
the service like so:

$ systemctl enable wpa_supplicant@{mlan0|wlan0}

And you should have a connection on the next boot.