gpsd: Allow gpsd to start on boot
The gpsd recipe creates a static systemd service. The original intention
was to make udev rule to invoke the gpsd  systemd interface, but it
currently has these two problems:

1. It only works with GPS devices connected over USB.
2. The recipe provides a very outdated configuration file
(gpsd.hotplug) that even with a USB GPS, it would not work.

This adds the start dependency and the [Install] section to the gpsd
service file provided in the recipe.

Signed-off-by: Ash Charles <ashcharles@gmail.com>