Add gumstix user in image, not via local.conf
A generic 'gumstix' user is added to the gumstix-console-image and
images which inherit from this recipe.  This demos adding basic user
accounts and provides a non-root user for both console and graphical
environments.  It is more explicit and versatile to add this user to
images rather than tucking it away in the local.conf.sample.

Users should update their local.conf to remove the addition of the
gumstix user.

Signed-off-by: Ash Charles <ashcharles@gmail.com>