wpa_supplicant: Provide default wlan0 configuration
Reorganize and provide a default configuration for wpa_supplicant. In
particular, boards with wifi will automatically connect to any
available unsecured wifi network on boot.  To change the network
settings (e.g. connect to a particular SSID secured by WPA2), edit the
wpa_supplicant.conf file.  To disable, either, on a board:
 $ systemctl disable wpa_supplicant@wlan0
or, in source, update the wpa-supplicant_%.bbappend file.

Signed-off-by: Ash Charles <ashcharles@gmail.com>