hostapd: Add hostapd to Gumstix images
Provide a basic hostapd configuration for users wanting to provide
a wireless access point with their Gumstix. This works for the SD8787
and Wilink8 chipsets (i.e. Pepper, DuoVero and Overo *-P COMs and
onward) but not those with the SD8686 chipset.  For these, see this [1]
guide.  To get started, try
 $ systemctl start hostapd
and watch for the 'gumstix' access point.  To enable this on boot:
 $ systemctl enable hostapd

[1] https://github.com/chrisw957/sd8686_uap

Signed-off-by: Ash Charles <ashcharles@gmail.com>