php: Remove phar from php package
As discussed [1], the "phar" utility is getting pulled into the php
package which, because host paths are leaking through, is causing an
invalid RDPENDS.  As such, just remove "phar" so the bad utility
doesn't get pulled in---a poor fix but it gets the job done until we
update to the corrected upstream version.  See the discussion here [2]
for more details.

[1] http://lists.openembedded.org/pipermail/openembedded-commits/2015-January/170064.html
[2] https://github.com/gumstix/Gumstix-YoctoProject-Repo/issues/59

Signed-off-by: Ash Charles <ashcharles@gmail.com>