systemd: DHCP Server should provide DNS server
As described [1], oblige systemd-network to respond with the server IP
as a DNS server when using the DHCPServer option.  When using access
point mode, it is no longer necessary to manually specify a DNS server
for each client---they should just work.

v2: supply Google's Public DNS at 8.8.8.8 as a DNS server is not
    necessarily available on the access point (in which case, this
    would all be moot).

[1] http://lists.freedesktop.org/archives/systemd-devel/2015-March/029848.html

Signed-off-by: Ash Charles <ashcharles@gmail.com>