dnsmasq: Use dnsmasq instead of systemd-networkd
To support wireless clients as an access point, the Gumstix board
should provide DHCP support including DNS---this is standard fare for
an access point.  Though nicely integrated, the systemd-networkd
DHCPServer option doesn't provide DNS for clients---we remove our patch
for systemd that provides upstream DNS.  Instead, use dnsmasq as the
standard lightweight solution providing a DHCP and DNS server.

It is automatically started by hostapd if this is used.

Signed-off-by: Ash Charles <ashcharles@gmail.com>