vim: Temporarily workaround hg fetcher error
The Fido branch introduces modification to bitbakes mercurial fetcher
that break existing downloaded tarballs.  For most users, the simple
solution is simply deleting existing tarballs for mercurial repository
donwloads.  This isn't possible on the autobuilder where pre-Fido
versions of the codebase are built against a shared downloads directory.
It is likewise not possible to override the behavior of the hg fetcher
from another layer.

Until this [1] is correctly resolved, work around by providing current
hashes for recipes relying on the hg fetcher.

This is temporary and fragile to tarballs getting updated.

[1] http://patchwork.openembedded.org/patch/94113/

Signed-off-by: Ash Charles <ashcharles@gmail.com>