scripts: Add EEPROM formatting utility
This script can be used to generate i2c commands for u-boot to write
board identification data to the board ID eeprom found on many
expansion boards. To generate EEPROM formatting commands
for a Gumstix Pepper43C, for example,
 $ uenv-generator -f R0011 -e '{"fdtfile": "am335x-pepper-43c.dtb"}' \
   -d 0x031 -r 0 -b 0 -a 0x50
The generated commands could then be pasted at the u-boot prompt or
included in a boot script.

Signed-off-by: Ash Charles <ashcharles@gmail.com>