Update to Krogoth : Rename and edit some bb files
network-manager-applet, libgles-oamp3, and vim bbappend files are renamed.

Also, recipe "midori" was used to build gumstix-xfce-image before. "midori" is removed from Krogoth and it is replaced by "epiphany"
mplayer2 which is used to build packagegroup-gumstix is blacklisted because of some compatibility issue with ffmpeg 3.
Finally, packagegroup-core-qt-demoapp and packagegroup-core-qt4e are removed because recipes-qt is removed in Krogoth.

Signed-off-by : Jerry Hung <jerry@gumstix.com>